Repro without a reboot
The only reason we needed a reboot is because the `environment.etc`
stuff happens *after* sops-nix, so the first boot/activation fails due
to a missing age key.

This gets the secret key into the vm using a nixos-shell mount, which
gets things to fail on the very first try.